% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCovidCohorts.R
\name{createCovidCohorts}
\alias{createCovidCohorts}
\title{Get the COVID related cohorts of the CHAPTER study}
\usage{
createCovidCohorts(cdm)
}
\arguments{
\item{cdm}{A cdm object as created using the
\code{\link[CDMConnector]{cdmFromCon}} function 
in the CDMConnector package.}
}
\description{
Get the COVID related cohorts of the CHAPTER study
}
\details{
This function applies the inclusion criteria for some cohorts which
must be created with bespoke code, and gets their attrition.
}
