% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenominatorCohort.R
\name{createDenominatorPopulation}
\alias{createDenominatorPopulation}
\title{Create a denominator population}
\usage{
createDenominatorPopulation(
  connectionDetails,
  cdmDatabaseSchema,
  targetDatabaseSchema,
  targetTable,
  oracleTempSchema = NULL,
  denominatorSettings
)
}
\description{
Create a denominator population
}
\details{
Create a denominator population by enf0orcing certain inclusion and exclusion criteria
}
