% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenominatorCohort.R
\name{createDenominatorSettings}
\alias{createDenominatorSettings}
\title{create the denominator population settings}
\usage{
createDenominatorSettings(
  washoutPeriod = 0,
  sampleSize = 0,
  useObservationPeriod = FALSE,
  cohortId = NULL,
  startDate = "",
  endDate = "",
  firstExposureOnly = F,
  requireDenominatorPeriod = F,
  minDenominatorPeriod = 0,
  denominatorDescription = ""
)
}
\arguments{
\item{washoutPeriod}{The mininum required continuous observation time (in days) prior to index
date for a person to be included in the cohort.}
}
\description{
create the denominator population settings
}
\details{
Takes as input the inputs to create denominator population
}
