% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line_length.R
\name{get_line_length}
\alias{get_line_length}
\title{Get line length}
\usage{
get_line_length(line.number = NULL)
}
\arguments{
\item{line.number}{A line number castable to integer}
}
\value{
Line lengt (int)
}
\description{
Get line length
}
