% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIncidenceResults.R
\name{getIncidenceResults}
\alias{getIncidenceResults}
\title{Get the incidence results of the CHAPTER study}
\usage{
getIncidenceResults(cdm, analyses)
}
\arguments{
\item{cdm}{A cdm object as created using the
\code{\link[CDMConnector]{cdmFromCon}} function 
in the CDMConnector package.}

\item{analyses}{A tibble specifying the incidence analyses to
conduct, retrieved from the settings folder}

\item{cohortNames}{Names of all the cohorts stored in the cdm}
}
\description{
Get the incidence results of the CHAPTER study
}
\details{
This function performs the incidence calculations and characterisation
of the relevant cohorts of the CHAPTER study.
}
