% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_last_char_same_delim.R
\name{is_last_char_same_delim}
\alias{is_last_char_same_delim}
\title{Check if last character is the used delimiter}
\usage{
is_last_char_same_delim(str, delim = NULL)
}
\arguments{
\item{str}{String to test}

\item{delim}{A delimiter character}
}
\value{
Logical
}
\description{
Check if last character is the used delimiter
}
