% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_last_char_whitespace.R
\name{is_last_char_whitespace}
\alias{is_last_char_whitespace}
\title{Check if last character is whitespace}
\usage{
is_last_char_whitespace(str, whitespace = c(" ", " \\t"))
}
\arguments{
\item{str}{String to test}

\item{whitespace}{A character vector defining whitespace}
}
\value{
Logical
}
\description{
Check if last character is whitespace
}
