% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line_length.R
\name{get_line_length}
\alias{get_line_length}
\title{Get line length}
\usage{
get_line_length(
  context = NULL,
  prim_selection = NULL,
  line.number = NULL,
  verbose = TRUE
)
}
\arguments{
\item{context}{Object of class 'document_context'}

\item{prim_selection}{A list, derived from context with rstudioapi::primary_selection()}

\item{line.number}{A line number castable to integer}

\item{verbose}{Print messages}
}
\value{
Line lengt (int)
}
\description{
Get line length
}
